+++ src/datasources/glofas.py
@@ -36,6 +36,19 @@ GF_STATIONS = {
 }
 
 
+def get_coords(station_name):
+    station = GF_STATIONS[station_name]
+    glofas_lon, glofas_lat = get_glofas_grid_coords(
+        station["lon"], station["lat"]
+    )
+    pitch = 0.001
+    N = glofas_lat + pitch
+    S = glofas_lat
+    E = glofas_lon + pitch
+    W = glofas_lon
+    return [N, W, S, E]
+
+
 def get_blob_name(
     data_type: Literal["raw", "processed"],
     dataset: Literal["reanalysis", "reforecast", "forecast"],
